// src/types/generated/nature-remo-openapi.ts
export interface paths {
// ...
	"/1/appliances": {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		/**
		 * @description Fetch the list of appliances.
		 *     Requires basic.read OAuth2 scopes.
		 */
		get: {
			parameters: {
				query?: never;
				header?: never;
				path?: never;
				cookie?: never;
			};
			requestBody?: never;
			responses: {
				/** @description ApplianceResponses */
				200: {
					headers: {
						[name: string]: unknown;
					};
					content: {
						"application/json": components["schemas"]["ApplianceResponses"];
					};
				};
			};
		};
		put?: never;
		/**
		 * @description Create a new appliance.
		 *     Requires basic.write OAuth2 scopes.
		 */
		post: {
			parameters: {
				query?: never;
				header?: never;
				path?: never;
				cookie?: never;
			};
			/** @description CreateApplianceRequest */
			requestBody: {
				content: {
					"application/x-www-form-urlencoded": components["schemas"]["CreateApplianceRequest"];
				};
			};
			responses: {
				/** @description ApplianceResponse */
				200: {
					headers: {
						[name: string]: unknown;
					};
					content: {
						"application/json": components["schemas"]["ApplianceResponse"];
					};
				};
			};
		};
		delete?: never;
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
// ...
	"/1/appliances/{applianceid}/aircon_settings": {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		get?: never;
		put?: never;
		/**
		 * @description Update air conditioner settings.
		 *     Requires sendir OAuth2 scopes.
		 */
		post: {
			parameters: {
				query?: never;
				header?: never;
				path: {
					/** @description Appliance Id */
					applianceid: string;
				};
				cookie?: never;
			};
			/** @description AirConParams_ */
			requestBody: {
				content: {
					"application/x-www-form-urlencoded": components["schemas"]["AirConParams_"];
				};
			};
			responses: {
				/** @description AC settings. */
				200: {
					headers: {
						[name: string]: unknown;
					};
					content: {
						"application/json": components["schemas"]["AirconSettingsResponse"];
					};
				};
			};
		};
		delete?: never;
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
// ...
	"/1/appliances/{applianceid}/light": {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		get?: never;
		put?: never;
		/**
		 * @description Send light signal.
		 *     Requires sendir OAuth2 scopes.
		 */
		post: {
			parameters: {
				query?: never;
				header?: never;
				path: {
					/** @description Appliance Id */
					applianceid: string;
				};
				cookie?: never;
			};
			/** @description Light parameters. */
			requestBody: {
				content: {
					"application/x-www-form-urlencoded": components["schemas"]["LightParams"];
				};
			};
			responses: {
				/** @description Light state. */
				200: {
					headers: {
						[name: string]: unknown;
					};
					content: {
						"application/json": components["schemas"]["LightState"];
					};
				};
			};
		};
		delete?: never;
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
// ...
	"/1/appliances/{applianceid}/tv": {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		get?: never;
		put?: never;
		/**
		 * @description Send TV signal.
		 *     Requires sendir OAuth2 scopes.
		 */
		post: {
			parameters: {
				query?: never;
				header?: never;
				path: {
					/** @description Appliance Id */
					applianceid: string;
				};
				cookie?: never;
			};
			/** @description TV parameters. */
			requestBody: {
				content: {
					"application/x-www-form-urlencoded": components["schemas"]["TVParams"];
				};
			};
			responses: {
				/** @description TV state. */
				200: {
					headers: {
						[name: string]: unknown;
					};
					content: {
						"application/json": components["schemas"]["TVState"];
					};
				};
			};
		};
		delete?: never;
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
// ...
	"/1/devices": {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		/**
		 * @description Fetch the list of Remo devices.
		 *     Requires basic.read OAuth2 scopes.
		 */
		get: {
			parameters: {
				query?: never;
				header?: never;
				path?: never;
				cookie?: never;
			};
			requestBody?: never;
			responses: {
				/** @description DeviceResponses */
				200: {
					headers: {
						[name: string]: unknown;
					};
					content: {
						"application/json": components["schemas"]["DeviceResponses"];
					};
				};
			};
		};
		put?: never;
		post?: never;
		delete?: never;
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
// ...
	"/1/signals/{signalid}/send": {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		get?: never;
		put?: never;
		/**
		 * @description Send a signal.
		 *     Requires sendir OAuth2 scopes.
		 */
		post: {
			parameters: {
				query?: never;
				header?: never;
				path: {
					/** @description Signal Id */
					signalid: string;
				};
				cookie?: never;
			};
			/** @description EmptyObject */
			requestBody: {
				content: {
					"application/x-www-form-urlencoded": components["schemas"]["EmptyObject"];
				};
			};
			responses: {
				/** @description EmptyObject */
				200: {
					headers: {
						[name: string]: unknown;
					};
					content: {
						"application/json": components["schemas"]["EmptyObject"];
					};
				};
			};
		};
		delete?: never;
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
// ...
}